+++ src/Sites/Commands/Workers/ListWorkersCommand.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace Laravel\Forge\Sites\Commands\Workers;
+
+class ListWorkersCommand extends WorkerCommand
+{
+    /**
+     * Items key for List response.
+     *
+     * @return string
+     */
+    public function listResponseItemsKey()
+    {
+        return 'workers';
+    }
+
+    /**
+     * HTTP request method.
+     *
+     * @return string
+     */
+    public function requestMethod()
+    {
+        return 'GET';
+    }
+}